refactor: split tools/invoker.py and tools/sandbox/docker_sandbox.py
